+++ src/main/java/games/rednblack/editor/renderer/components/LayoutComponent.java
@@ -13,7 +13,7 @@ public class LayoutComponent extends PooledComponent {
     public float horizontalBias = 0.5f;
     public float verticalBias = 0.5f;
 
-    public transient float checksum;
+    public transient int checksum;
 
     public static class ConstraintData {
         @EntityId public int targetEntity = -1; // -1 = parent
